legal sys/requirements.txt
flask
flask-cors
sentence-transformers
numpy
google-generativeai
python-dotenv
